backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
yfinance>=0.2.36
google-genai>=0.2.2
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.5
ruff>=0.2.1
